+++ docs/api/examples/eth/ETH_Sender_Receiver.cpp
@@ -21,21 +21,29 @@ ethernetSender->Activate();
 
 // ------------------------------------------------------------
 // Send an Ethernet message
-std::array<uint8_t, 6> sourceAddress{"F6", "04", "68", "71", "AA", "C1"};
-std::array<uint8_t, 6> destinationAddress{"F6", "04", "68", "71", "AA", "C2"};
+const std::array<uint8_t, 6> destinationAddress{0xf6, 0x04, 0x68, 0x71, 0xaa, 0xc2};
+const std::array<uint8_t, 6> sourceAddress{0xf6, 0x04, 0x68, 0x71, 0xaa, 0xc1};
+const std::array<uint8_t, 4> vlanTag{0x81, 0x00, 0x00, 0x00};  // optional
+const std::array<uint8_t, 2> etherType{0x08, 0x00};
+const std::array<uint8_t, 4> frameCheckSequence{0x00, 0x00, 0x00, 0x00};  // optional
 
-std::string message{"Ensure that the payload is long enough to constitute "
+const std::string message{"Ensure that the payload is at least 46 bytes to constitute "
                     "a valid Ethernet frame ------------------------------"};
-std::vector<uint8_t> payload{message.begin(), message.end()};
+const std::vector<uint8_t> payload{message.begin(), message.end()};
 
-EthernetFrame ethFrame;
-ethFrame.SetSourceMac(sourceAddress);
-ethFrame.SetDestinationMac(destinationAddress);
-ethFrame.SetPayload(payload);
+const std::array<uint8_t, 2> frameCheckSequence{0x00, 0x00};  // optional for 
 
-// The returned transmitId can be used to check if the ethTransmitAcknowledge
+EthernetFrame frame{};
+std::copy(destinationAddress.begin(), destinationAddress.end(), std::back_inserter(frame));
+std::copy(sourceAddress.begin(), sourceAddress.end(), std::back_inserter(frame));
+std::copy(vlanTag.begin(), vlanTag.end(), std::back_inserter(frame));
+std::copy(etherType.begin(), etherType.end(), std::back_inserter(frame));
+std::copy(payload.begin(), payload.end(), std::back_inserter(frame));
+std::copy(frameCheckSequence.begin(), frameCheckSequence.end(), std::back_inserter(frame));
+
+// The returned transmitId can be used to check if the ethernetFrameTransmitEvent
 // that should be triggered after a successful reception has the same transmitId.
-auto transmitId = ethernetSender->SendFrame(ethFrame);
+auto transmitId = ethernetSender->SendFrame(frame);
 
 
 // ------------------------------------------------------------
@@ -43,10 +51,10 @@ auto transmitId = ethernetSender->SendFrame(ethFrame);
 //  - TX confirmation for the sender.
 sender_FrameTransmitHandler(ethernetSender, ethernetFrameTransmitEvent);
 // with:
-//  - ethTransmitAcknowledge.transmitId == 1
-//  - ethTransmitAcknowledge.sourceMac == {"F6", "04", "68", "71", "AA", "C1"}
-//  - ethTransmitAcknowledge.timestamp == <Timestamp of EthernetFrame>
-//  - ethTransmitAcknowledge.status == EthernetTransmitStatus::Transmitted 
+//  - ethernetFrameTransmitEvent.transmitId == 1
+//  - ethernetFrameTransmitEvent.sourceMac == {0xF6, 0x04, 0x68, 0x71, 0xAA, 0xC1}
+//  - ethernetFrameTransmitEvent.timestamp == <Timestamp of EthernetFrame>
+//  - ethernetFrameTransmitEvent.status == EthernetTransmitStatus::Transmitted 
 // Note: When using the VIBE NetworkSimulator, the status can also be EthernetTransmitStatus::LinkDown.
 
 //  - RX Ethernet message for the receiver.